# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "GeoClimate": {
        "additionalProperties": false,
        "description": "The climate zone a point resolved to, and the class it came from.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "koppen_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Koppen Class"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "resolution_caveat": {
            "title": "Resolution Caveat",
            "type": "string"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "zone",
          "koppen_class",
          "source_key",
          "attribution",
          "note",
          "resolution_caveat"
        ],
        "title": "GeoClimate",
        "type": "object"
      },
      "GeoCountry": {
        "additionalProperties": false,
        "description": "The country a point resolved to, and how the lookup reached it.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "iso_a2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iso A2"
          },
          "iso_a3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iso A3"
          },
          "matched": {
            "title": "Matched",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          }
        },
        "required": [
          "iso_a2",
          "iso_a3",
          "name",
          "matched",
          "source_key",
          "attribution"
        ],
        "title": "GeoCountry",
        "type": "object"
      },
      "GeoLookupRequest": {
        "additionalProperties": false,
        "description": "``POST /api/geo/lookup`` — what a map click can honestly answer (D-047).\n\n``boundary`` is the polygon the user drew, as ``[longitude, latitude]``\npairs. When it is absent the user placed a point rather than drawing a\nsite, and no area is returned.",
        "properties": {
          "boundary": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "GeoLookupRequest",
        "type": "object"
      },
      "GeoLookupResponse": {
        "additionalProperties": false,
        "description": "``POST /api/geo/lookup``.\n\nEvery value is a suggestion. The interface applies each one only where the\nuser has not already answered, marks what it applied as autofilled, and\nlets any of them be overridden (D-047.2).",
        "properties": {
          "climate": {
            "$ref": "#/components/schemas/GeoClimate"
          },
          "country": {
            "$ref": "#/components/schemas/GeoCountry"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "site_area_m2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Area M2"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "site_area_m2",
          "country",
          "climate"
        ],
        "title": "GeoLookupResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceResult": {
        "additionalProperties": false,
        "description": "One place-search match (v2.2, D-049.6): a name and somewhere to move\nthe map to. Never an answer to any questionnaire field.",
        "properties": {
          "admin": {
            "title": "Admin",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "population": {
            "title": "Population",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "admin",
          "latitude",
          "longitude",
          "population"
        ],
        "title": "PlaceResult",
        "type": "object"
      },
      "PlaceSearchResponse": {
        "additionalProperties": false,
        "description": "``GET /api/geo/places`` — offline navigation by name.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PlaceResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results",
          "attribution"
        ],
        "title": "PlaceSearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Nature for Cooling Rapid Assessment Tool",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/geo/lookup": {
      "post": {
        "description": "Answer the three questions a map may answer about a location.\n\nReturns suggestions, not answers. The caller applies each one only where\nthe user has not already answered, and marks what it applies as autofilled\n(D-047.2) — an autofilled value never overwrites an answer already given.",
        "operationId": "lookup_api_geo_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeoLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoLookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lookup",
        "tags": [
          "geo"
        ]
      }
    },
    "/api/geo/places": {
      "get": {
        "description": "Search the bundled populated-places index by name (v2.2, D-049.6).\n\nNavigation, not autofill: a result is somewhere for the map to move to,\nand selecting one fills in no answer. Answered entirely from bundled\npublic-domain data — this is the navigation aid a deployment without\nconfigured imagery has.",
        "operationId": "places_api_geo_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Places",
        "tags": [
          "geo"
        ]
      }
    },
...
  }
}